requirements.txt
asgiref==3.7.2
asttokens==2.4.1
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
ipykernel==6.27.0
ipython==8.18.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
libtmux==0.24.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.12
